refactor(test): move resuable funcs, add concurrency helpers
