+++ cmd/concurrency_helpers_test.go
@@ -1,0 +1,5 @@
+package cmd
+
+import "github.com/grafvonb/c8volt/testx"
+
+type safeSlice[T any] = testx.SafeSlice[T]
