+++ testx/auth_oauth2_response.go
@@ -4,7 +4,6 @@ import (
 	"net/http"
 	"testing"
 
-	"github.com/grafvonb/c8volt/config"
 	"github.com/grafvonb/c8volt/internal/clients/auth/oauth2"
 )
 
@@ -26,37 +25,5 @@ func TestAuthJSON200Response(t *testing.T, status int, token string, raw string)
 			TokenType:   "Bearer",
 		},
 		HTTPResponse: &http.Response{StatusCode: status},
-	}
-}
-
-func TestConfig(t *testing.T) *config.Config {
-	t.Helper()
-	return &config.Config{
-		App: config.App{
-			Tenant: "tenant",
-		},
-		Auth: config.Auth{
-			OAuth2: config.AuthOAuth2ClientCredentials{
-				TokenURL:     "http://localhost/token",
-				ClientID:     "test",
-				ClientSecret: "test",
-			},
-			Cookie: config.AuthCookieSession{
-				BaseURL:  "http://localhost/cookie",
-				Username: "test",
-				Password: "test",
-			},
-		},
-		APIs: config.APIs{
-			Camunda: config.API{
-				BaseURL: "http://localhost/camunda/v2",
-			},
-			Operate: config.API{
-				BaseURL: "http://localhost/operate",
-			},
-			Tasklist: config.API{
-				BaseURL: "http://localhost/tasklist",
-			},
-		},
 	}
 }
